requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
python-docx>=0.8.11
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
python-dateutil>=2.8.0
aiohttp>=3.8.0
